requirements.txt
Django==1.7.1
djangorestframework==3.0.4
django-post-office==1.1.1